Clear up confusion about event in documentation
It happens so often to me, that when I want to install socket.io, that I listen for `event` and get nothing. Just as I dig deeper into the documentation and discover that `event` is actually the name of the event. Not that I can actually listen on `event` and get any data.

So to clear that common misconception up, I would add a little comment to tell what `event` actually is. As `connect` and `disconnect` are pretty self-explanatory. However, if you see `event` between the two, it seems like it really has to be `event`. As if there are just three events to listen on: connect, event and disconnect.

If you are not happy with the wording, please let me know